cdk/requirements.txt
annotated-types==0.6.0
attrs==23.2.0
aws-cdk-lib==2.132.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
cattrs==23.2.3
constructs==10.3.0
importlib_resources==6.3.1
jsii==1.95.0
publication==0.0.3
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.10.0
